Implement timer for menu system

+++ node1/usb_multifunction_card/menu.h
@@ -3,5 +3,5 @@
 
 void menu_draw_options(const char ** options, uint8_t n_options);
 int menu_select_option(uint8_t n_options);
-void menu_display_score(uint8_t score);
+void menu_display_score(uint16_t score);
 #endif
